sooo...many...tests

+++ pandas/tests/frame/methods/test_asfreq.py
@@ -159,7 +159,7 @@ class TestAsFreq:
 
         # setup
         rng = date_range("1/1/2016", periods=10, freq="2S")
-        ts = Series(np.arange(len(rng)), index=rng)
+        ts = Series(np.arange(len(rng)), index=rng, dtype=float)
         df = DataFrame({"one": ts})
 
         # insert pre-existing missing value
